pyproject.toml
[tool.poetry]
name = "tuw-master-thesis"
version = "0.1.0"
description = ""
authors = ["Jan Hadl <jan.d.hadl@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
clingo = "^5.6.2"
tensorflow = { version = "^2.11.0", markers = "sys_platform != 'darwin'"}
tensorflow-macos = { version = "^2.11.0", markers = "sys_platform == 'darwin'"}
tensorflow-metal = { version = "^0.7.1", markers = "sys_platform == 'darwin'"}
transformers = "^4.26.1"
pillow = "^9.4.0"
requests = "^2.28.2"
torch = "^2.0.0"
torchvision = "^0.15.1"
timm = "^0.6.12"
matplotlib = "^3.7.1"
nltk = "^3.8.1"
pandas = "^1.5.3"
sentence-transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
jupyter-client = "7.3.2"
tornado = "6.1"
notebook = "^6.5.2"
ipywidgets = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
